requirements.txt
streamlit>=1.25.0
ultralytics>=8.0.146
opencv-python
Pillow
numpy
torch>=2.0.0
